Destructible Buildings
Added a new Level (BuildingTests).

Added a ton of building sprite assets.

Added a blueprint for destructible buildings (BP_Building_2 - BP_Building is deprecated).
Blueprint sprites/flipbooks can be swapped out per-instance. Collision updates correctly.

This level has a few buildings in the scene and an object called TestItem_ClickForExplosions - when in a scene, this will override the player controller for a test mode:
Click anywhere to trigger an explosion. Enough damage to a building will destroy it. Destroyed buildings play a flipbook, spawn a remnant sprite (if one exists) and create an explosion and smoke emitter (placeholders)

Bugs:
- explosions on click is not exact, but this is just for testing.
- if a building takes damage while the flipbook is playing, it retriggers.